% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{kmeans_stream_predict}
\alias{kmeans_stream_predict}
\title{Predict Cluster Assignment for One Observation}
\usage{
kmeans_stream_predict(model, x)
}
\arguments{
\item{model}{a streaming k-means model created by
\code{\link{kmeans_stream_new}}}

\item{x}{a numeric vector of length \code{d}}
}
\value{
integer, the cluster assignment for \code{x} (1-indexed)
}
\description{
Returns the index of the nearest centroid to \code{x} without
updating the model.
}
\examples{
model <- kmeans_stream_new(k = 2L, d = 2L)
kmeans_stream_update(model, c(0.0, 0.0))
kmeans_stream_update(model, c(5.0, 5.0))
kmeans_stream_predict(model, c(0.1, 0.1))

}
\seealso{
\code{\link{kmeans_stream_new}},
\code{\link{kmeans_stream_update}}
}
