% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stReans-package.R
\docType{package}
\name{stReans-package}
\alias{stReans}
\alias{stReans-package}
\title{stReans: Batch and Streaming K-Means Clustering with a C Backend}
\description{
Implements Lloyd's algorithm for batch data and an incremental online k-means for streaming data. The core computations are performed in C via the R native API for performance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Luki308/stReans}
  \item \url{https://luki308.github.io/stReans/}
}

}
\author{
\strong{Maintainer}: Łukasz Lepianka \email{01159348@pw.edu.pl}

Authors:
\itemize{
  \item Łukasz Lepianka \email{01159348@pw.edu.pl}
}

}
\keyword{internal}
